Added info router, and db class
