+++ src/index.js
@@ -2,6 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const errorHandler = require('./middleware/errorHandler');
 const apiRouter = require('./routers/apiRouter');
+const infoRouter = require('./routers/infoRouter');
 
 //Server setup
 const port = process.env.PORT || 3000;
@@ -11,6 +12,7 @@ app.use(express.json());
 
 //Api Path
 app.use('/api', apiRouter);
+app.use('/info', infoRouter);
 
 //Setup error handler
 app.use(errorHandler);
