+++ src/routers/infoRouter.js
@@ -1,0 +1,4 @@
+const express = require('express');
+const router = express.Router();
+
+module.exports = router;
